+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/informers/SharedInformerFactory.java
@@ -25,17 +25,15 @@ import io.fabric8.kubernetes.api.model.ListOptions;
 import io.fabric8.kubernetes.client.Config;
 import io.fabric8.kubernetes.client.ConfigBuilder;
 import io.fabric8.kubernetes.client.CustomResource;
-import io.fabric8.kubernetes.client.Handlers;
-import io.fabric8.kubernetes.client.ResourceHandler;
 import io.fabric8.kubernetes.client.SharedInformerOperationsImpl;
 import io.fabric8.kubernetes.client.Watch;
 import io.fabric8.kubernetes.client.Watcher;
+import io.fabric8.kubernetes.client.dsl.Informable;
 import io.fabric8.kubernetes.client.dsl.Resource;
 import io.fabric8.kubernetes.client.dsl.base.BaseOperation;
 import io.fabric8.kubernetes.client.dsl.base.CustomResourceDefinitionContext;
 import io.fabric8.kubernetes.client.dsl.base.OperationContext;
 import io.fabric8.kubernetes.client.informers.impl.DefaultSharedIndexInformer;
-import io.fabric8.kubernetes.client.utils.ApiVersionUtil;
 import io.fabric8.kubernetes.client.utils.Utils;
 import io.fabric8.kubernetes.internal.KubernetesDeserializer;
 
@@ -59,7 +57,7 @@ import okhttp3.OkHttpClient;
  * This has been taken from https://github.com/kubernetes-client/java/blob/master/util/src/main/java/io/kubernetes/client/informer/SharedInformerFactory.java
  * which is ported from offical go client https://github.com/kubernetes/client-go/blob/master/informers/factory.go
  */
-public class SharedInformerFactory extends BaseOperation {
+public class SharedInformerFactory {
   private final List<Map.Entry<OperationContext, SharedIndexInformer>> informers = new ArrayList<>();
 
   private final ExecutorService informerExecutor;
@@ -67,6 +65,28 @@ public class SharedInformerFactory extends BaseOperation {
   private final ConcurrentLinkedQueue<SharedInformerEventListener> eventListeners = new ConcurrentLinkedQueue<>();
 
   private boolean allowShutdown = true;
+  
+  private class SharedInformerFactoryBaseOperation extends BaseOperation {
+
+    private SharedInformerFactoryBaseOperation(OperationContext ctx) {
+      super(ctx);
+    }
+    
+    private OperationContext getContext() {
+      return this.context;
+    }
+    
+    private void setContext(OperationContext context) {
+      this.context = context;
+    }
+    
+    private void setName(String name) {
+      this.name = name;
+    }
+    
+  }
+  
+  private SharedInformerFactoryBaseOperation baseOperation;
   
   public SharedInformerFactory(OkHttpClient okHttpClient, Config configuration) {
     // ideally this should be bounded.  The current implication is that there
@@ -84,10 +104,10 @@ public class SharedInformerFactory extends BaseOperation {
    * @param configuration configuration for client
    */
   public SharedInformerFactory(ExecutorService threadPool, OkHttpClient okHttpClient, Config configuration) {
-    super(new OperationContext().withOkhttpClient(okHttpClient).withConfig(configuration));
+    this.baseOperation = new SharedInformerFactoryBaseOperation(new OperationContext().withOkhttpClient(okHttpClient).withConfig(configuration));
     initOperationContext(configuration);
     this.informerExecutor = threadPool;
-    this.namespace = null;
+    this.baseOperation.setNamespace(null);
   }
 
   /**
@@ -95,10 +115,11 @@ public class SharedInformerFactory extends BaseOperation {
    *
    * @param namespace namespace to configure
    * @return {@link SharedInformerFactory} with namespace configured
+   * @deprecated use {@link Informable} instead
    */
-  @Override
+  @Deprecated
   public SharedInformerFactory inNamespace(String namespace) {
-    this.namespace = namespace;
+    this.baseOperation.setNamespace(namespace);
     return this;
   }
 
@@ -107,10 +128,11 @@ public class SharedInformerFactory extends BaseOperation {
    *
    * @param name name to be configured
    * @return {@link SharedInformerFactory} with name configured
+   * @deprecated use {@link Informable} instead
    */
-  @Override
+  @Deprecated
   public SharedInformerFactory withName(String name) {
-    this.name = name;
+    this.baseOperation.setName(name);
     return this;
   }
 
@@ -139,7 +161,9 @@ public class SharedInformerFactory extends BaseOperation {
    * @param resyncPeriodInMillis resync period in milliseconds
    * @param <T> the type parameter (should extend {@link io.fabric8.kubernetes.api.model.HasMetadata} and implement {@link io.fabric8.kubernetes.api.model.Namespaced}) if Namespace scoped resource
    * @return the shared index informer
+   * @deprecated use {@link Informable} instead
    */
+  @Deprecated
   public synchronized <T extends HasMetadata> SharedIndexInformer<T> sharedIndexInformerFor(Class<T> apiTypeClass, OperationContext operationContext, long resyncPeriodInMillis) {
     return sharedIndexInformerFor(apiTypeClass, inferListType(apiTypeClass), operationContext, resyncPeriodInMillis, Utils.isResourceNamespaced(apiTypeClass));
   }
@@ -165,11 +189,11 @@ public class SharedInformerFactory extends BaseOperation {
       Objects.requireNonNull(customResourceContext.getGroup()),
       Objects.requireNonNull(customResourceContext.getVersion())
     ), Optional.ofNullable(customResourceContext.getKind()).orElse(apiTypeClass.getSimpleName()), apiTypeClass);
-    return sharedIndexInformerFor(apiTypeClass, apiListTypeClass, context
+    return sharedIndexInformerFor(apiTypeClass, apiListTypeClass, baseOperation.getContext()
       .withApiGroupName(customResourceContext.getGroup())
       .withApiGroupVersion(customResourceContext.getVersion())
       .withPlural(customResourceContext.getPlural())
-      .withIsNamespaceConfiguredFromGlobalConfig(context.isNamespaceFromGlobalConfig()), resyncPeriodInMillis, Utils.isResourceNamespaced(apiTypeClass));
+      .withIsNamespaceConfiguredFromGlobalConfig(baseOperation.getContext().isNamespaceFromGlobalConfig()), resyncPeriodInMillis, Utils.isResourceNamespaced(apiTypeClass));
   }
 
   /**
@@ -184,13 +208,12 @@ public class SharedInformerFactory extends BaseOperation {
    * @return {@link SharedIndexInformer} for GenericKubernetesResource
    */
   public synchronized SharedIndexInformer<GenericKubernetesResource> sharedIndexInformerForCustomResource(CustomResourceDefinitionContext genericResourceContext, long resyncPeriodInMillis) {
-    validateCustomResourceProvided(genericResourceContext);
     return sharedIndexInformerFor(GenericKubernetesResource.class, GenericKubernetesResourceList.class, new OperationContext()
       .withApiGroupName(genericResourceContext.getGroup())
       .withApiGroupVersion(genericResourceContext.getVersion())
       .withPlural(genericResourceContext.getPlural())
-      .withNamespace(this.namespace)
-      .withIsNamespaceConfiguredFromGlobalConfig(context.isNamespaceFromGlobalConfig()), resyncPeriodInMillis, genericResourceContext.isNamespaceScoped());
+      .withNamespace(baseOperation.getNamespace())
+      .withIsNamespaceConfiguredFromGlobalConfig(baseOperation.getContext().isNamespaceFromGlobalConfig()), resyncPeriodInMillis, genericResourceContext.isNamespaceScoped());
   }
 
   /**
@@ -200,7 +223,9 @@ public class SharedInformerFactory extends BaseOperation {
    * @param resyncPeriodInMillis resync period in milliseconds
    * @param <T> the type parameter (should extend {@link CustomResource} and implement {@link io.fabric8.kubernetes.api.model.Namespaced})
    * @return the shared index informer
+   * @deprecated use {@link Informable} instead
    */
+  @Deprecated
   public synchronized <T extends CustomResource<?,?>> SharedIndexInformer<T> sharedIndexInformerForCustomResource(
     Class<T> apiTypeClass, OperationContext operationContext, long resyncPeriodInMillis) {
     return sharedIndexInformerFor(apiTypeClass, inferListType(apiTypeClass), operationContext, resyncPeriodInMillis, Utils.isResourceNamespaced(apiTypeClass));
@@ -251,15 +276,15 @@ public class SharedInformerFactory extends BaseOperation {
    */
   private synchronized <T extends HasMetadata, L extends KubernetesResourceList<T>> SharedIndexInformer<T> sharedIndexInformerFor(Class<T> apiTypeClass, Class<L> apiListTypeClass, OperationContext operationContext, long resyncPeriodInMillis, boolean isResourceNamespaced) {
     ListerWatcher<T, L> listerWatcher = listerWatcherFor(apiTypeClass, apiListTypeClass, isResourceNamespaced);
-    OperationContext context = this.context.withApiGroupName(HasMetadata.getGroup(apiTypeClass))
+    OperationContext context = this.baseOperation.getContext().withApiGroupName(HasMetadata.getGroup(apiTypeClass))
       .withApiGroupVersion(HasMetadata.getVersion(apiTypeClass))
       .withPlural(HasMetadata.getPlural(apiTypeClass))
-      .withIsNamespaceConfiguredFromGlobalConfig(this.context.isNamespaceFromGlobalConfig());
-    if (this.namespace != null) {
-      context = context.withNamespace(this.namespace).withIsNamespaceConfiguredFromGlobalConfig(false);
+      .withIsNamespaceConfiguredFromGlobalConfig(this.baseOperation.getContext().isNamespaceFromGlobalConfig());
+    if (this.baseOperation.getNamespace() != null) {
+      context = context.withNamespace(this.baseOperation.getNamespace()).withIsNamespaceConfiguredFromGlobalConfig(false);
     }
-    if (this.name != null) {
-      context = context.withFields(Collections.singletonMap("metadata.name", this.name));
+    if (this.baseOperation.getName() != null) {
+      context = context.withFields(Collections.singletonMap("metadata.name", this.baseOperation.getName()));
     }
     if (operationContext != null) {
       context = context.withOperationContext(operationContext);
@@ -315,7 +340,9 @@ public class SharedInformerFactory extends BaseOperation {
    * by the user.
    *
    * @return a list of entries of {@link OperationContext} and {@link SharedIndexInformer}
+   * @deprecated use {@link #getExistingSharedIndexInformer(Class)} instead
    */
+  @Deprecated
   public List<Map.Entry<OperationContext, SharedIndexInformer>> getExistingSharedIndexInformers() {
     return Collections.unmodifiableList(this.informers);
   }
@@ -349,9 +376,6 @@ public class SharedInformerFactory extends BaseOperation {
    * @param shutDownThreadPool Whether to shut down thread pool or not.
    */
   public synchronized void stopAllRegisteredInformers(boolean shutDownThreadPool) {
-    if (informers.isEmpty()) {
-      return;
-    }
     informers.forEach(e -> e.getValue().stop());
     if (shutDownThreadPool && allowShutdown) {
       informerExecutor.shutdown();
@@ -383,7 +407,7 @@ public class SharedInformerFactory extends BaseOperation {
       // SharedInformer default behavior is to watch in all namespaces
       // unless we specify namespace explicitly in OperationContext
       sharedInformerOperations = new SharedInformerOperationsImpl<>(context
-        .withConfig(new ConfigBuilder(config)
+        .withConfig(new ConfigBuilder(this.baseOperation.getConfig())
           .withNamespace(null)
           .build())
         .withNamespace(null), isNamespacedScoped);
@@ -397,7 +421,7 @@ public class SharedInformerFactory extends BaseOperation {
 
   private void initOperationContext(Config configuration) {
     if (configuration.getNamespace() != null) {
-      context = context.withIsNamespaceConfiguredFromGlobalConfig(true);
+      this.baseOperation.setContext(this.baseOperation.getContext().withIsNamespaceConfiguredFromGlobalConfig(true));
     }
   }
 
@@ -407,14 +431,4 @@ public class SharedInformerFactory extends BaseOperation {
     }
   }
 
-  private void validateCustomResourceProvided(CustomResourceDefinitionContext genericResourceContext) {
-    if (Utils.isNullOrEmpty(genericResourceContext.getKind())) {
-      throw new IllegalArgumentException("CustomResourceDefinitionContext.kind: required value.");
-    }
-    ResourceHandler<HasMetadata, ?> resourceHandler = Handlers.get(genericResourceContext.getKind(), ApiVersionUtil.joinApiGroupAndVersion(genericResourceContext.getGroup(), genericResourceContext.getVersion()));
-
-    if (resourceHandler != null) {
-      throw new IllegalArgumentException("Using sharedIndexInformerDynamicResource for core type. Please use sharedIndexInformerFor(Class<T>, long) instead.");
-    }
-  }
 }